claude_agent_course/module_03_custom_tools/projects/project_02_weather_assistant/pyproject.toml
[project]
name = "weather-assistant"
version = "0.1.0"
description = "天气查询助手 - 自定义工具实战"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
